use int32 to identify users and items

+++ base/index.go
@@ -21,10 +21,10 @@ import (
 
 // Index keeps the mapping between names (string) and indices (integer).
 type Index interface {
-	Len() int
+	Len() int32
 	Add(name string)
-	ToNumber(name string) int
-	ToName(index int) string
+	ToNumber(name string) int32
+	ToName(index int32) string
 	GetNames() []string
 }
 
@@ -37,39 +37,39 @@ func init() {
 // a user ID or item ID. The dense index is the internal user index or item index
 // optimized for faster parameter access and less memory usage.
 type MapIndex struct {
-	Numbers map[string]int // sparse ID -> dense index
-	Names   []string       // dense index -> sparse ID
+	Numbers map[string]int32 // sparse ID -> dense index
+	Names   []string         // dense index -> sparse ID
 }
 
 // NotId represents an ID doesn't exist.
-const NotId = -1
+const NotId = int32(-1)
 
 // NewMapIndex creates a MapIndex.
 func NewMapIndex() *MapIndex {
 	set := new(MapIndex)
-	set.Numbers = make(map[string]int)
+	set.Numbers = make(map[string]int32)
 	set.Names = make([]string, 0)
 	return set
 }
 
 // Len returns the number of indexed Names.
-func (idx *MapIndex) Len() int {
+func (idx *MapIndex) Len() int32 {
 	if idx == nil {
 		return 0
 	}
-	return len(idx.Names)
+	return int32(len(idx.Names))
 }
 
 // Add adds a new ID to the indexer.
 func (idx *MapIndex) Add(name string) {
 	if _, exist := idx.Numbers[name]; !exist {
-		idx.Numbers[name] = len(idx.Names)
+		idx.Numbers[name] = int32(len(idx.Names))
 		idx.Names = append(idx.Names, name)
 	}
 }
 
 // ToNumber converts a sparse ID to a dense index.
-func (idx *MapIndex) ToNumber(name string) int {
+func (idx *MapIndex) ToNumber(name string) int32 {
 	if denseId, exist := idx.Numbers[name]; exist {
 		return denseId
 	}
@@ -77,7 +77,7 @@ func (idx *MapIndex) ToNumber(name string) int {
 }
 
 // ToName converts a dense index to a sparse ID.
-func (idx *MapIndex) ToName(index int) string {
+func (idx *MapIndex) ToName(index int32) string {
 	return idx.Names[index]
 }
 
@@ -89,7 +89,7 @@ func (idx *MapIndex) GetNames() []string {
 // DirectIndex means that the name and its index is the same. For example,
 // the index of "1" is 1, vice versa.
 type DirectIndex struct {
-	Limit int
+	Limit int32
 }
 
 // NewDirectIndex create a direct mapping index.
@@ -98,7 +98,7 @@ func NewDirectIndex() *DirectIndex {
 }
 
 // Len returns the number of names in current index.
-func (idx *DirectIndex) Len() int {
+func (idx *DirectIndex) Len() int32 {
 	return idx.Limit
 }
 
@@ -108,29 +108,29 @@ func (idx *DirectIndex) Add(s string) {
 	if err != nil {
 		panic(err)
 	}
-	if i >= idx.Limit {
-		idx.Limit = i + 1
+	if int32(i) >= idx.Limit {
+		idx.Limit = int32(i + 1)
 	}
 }
 
 // ToNumber converts a name to corresponding index.
-func (idx *DirectIndex) ToNumber(name string) int {
+func (idx *DirectIndex) ToNumber(name string) int32 {
 	i, err := strconv.Atoi(name)
 	if err != nil {
 		panic(err)
 	}
-	if i >= idx.Limit {
+	if int32(i) >= idx.Limit {
 		return NotId
 	}
-	return i
+	return int32(i)
 }
 
 // ToName converts a index to corresponding name.
-func (idx *DirectIndex) ToName(index int) string {
+func (idx *DirectIndex) ToName(index int32) string {
 	if index >= idx.Limit {
 		panic("index out of range")
 	}
-	return strconv.Itoa(index)
+	return strconv.Itoa(int(index))
 }
 
 // GetNames returns all names in current index.
